GCS: Quiet down some compiler warnings.
There are now 0 build issues in Linux when I use "QMAKE_CXXFLAGS+=-Wno-ignored-qualifiers -Wno-unused-parameter -Wno-unused-variable" (including the "") as additional arguments to qmake in Qt Creator.

git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@1232 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ ground/src/plugins/rawhid/pjrc_rawhid_win.cpp
@@ -209,7 +209,7 @@ int pjrc_rawhid::receive(int num, void *buf, int len, int timeout)
 
 //        qDebug("Received %i bytes, first %x, second %x", len, *((char *) buf),*((char *)buf + 1));
 
-        if (n > len) n = len;
+        if ((int)n > len) n = len;
         return n;
 return_timeout:
         CancelIo(hid->handle);
